# API-contract.yaml
openapi: 3.0.3
info:
  title: Function Storage API
  version: 1.0.0
  description: |
    Единый унифицированный контракт API для работы с табулированными функциями.
    Поддерживает:
      - операции CRUD и поиск по владельцу;
      - арифметические операции над функциями (сложение, вычитание, умножение, деление);
      - численное дифференцирование табулированных функций;
      - импорт/экспорт в текстовом и бинарном формате;
      - сериализацию/десериализацию функций.
servers:
  - url: http://localhost:8080/api
    description: |
      Local development (manual: Tomcat, framework: Spring Boot)
security:
  - BasicAuth: []
paths:
  /search/users/single/{username}:
    get:
      summary: Найти одного пользователя по имени
      tags: [ Search ]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: Имя пользователя для поиска
      responses:
        '200':
          description: Найденный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse' # Предполагаем наличие этой схемы
        '404':
          description: Пользователь не найден
  /search/users/multiple/role/{role}:
    get:
      summary: Найти пользователей по роли
      tags: [ Search ]
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
            enum: [ USER, ADMIN, MODERATOR ] # Используем Enum Role
          description: Роль для фильтрации
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Неверный формат роли
  /search/functions/by-user/{userId}:
    get:
      summary: Найти функции по ID пользователя
      tags: [ Search ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID владельца функции
      responses:
        '200':
          description: Список функций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FunctionListResponse' # Соответствует TabulatedFunctionListDTO
  /search/functions/by-name:
    get:
      summary: Найти функции, имя которых содержит подстроку
      tags: [ Search ]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Подстрока для поиска в имени функции
      responses:
        '200':
          description: Список функций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FunctionListResponse'
  /search/functions/with-min-points/{minPoints}:
    get:
      summary: Найти функции с минимальным количеством точек
      tags: [ Search ]
      parameters:
        - name: minPoints
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
          description: Минимальное количество точек
      responses:
        '200':
          description: Список функций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FunctionListResponse'
  /search/points/by-function/{functionId}:
    get:
      summary: Найти точки по ID функции
      tags: [ Search ]
      parameters:
        - name: functionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID функции, которой принадлежат точки
      responses:
        '200':
          description: Список точек
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointResponse' # Соответствует PointDTO
  /search/points/in-range:
    get:
      summary: Найти точки в заданном диапазоне X и Y
      tags: [ Search ]
      parameters:
        - name: minX
          in: query
          required: true
          schema: { type: number, format: double }
        - name: maxX
          in: query
          required: true
          schema: { type: number, format: double }
        - name: minY
          in: query
          required: true
          schema: { type: number, format: double }
        - name: maxY
          in: query
          required: true
          schema: { type: number, format: double }
      responses:
        '200':
          description: Список точек
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointResponse'
  /search/points/advanced:
    get:
      summary: Расширенный поиск точек с фильтрацией и сортировкой
      tags: [ Search ]
      parameters:
        - name: functionId
          in: query
          required: false
          schema: { type: integer, format: int64 }
        - name: minX
          in: query
          required: false
          schema: { type: number, format: double }
        - name: maxX
          in: query
          required: false
          schema: { type: number, format: double }
        - name: minY
          in: query
          required: false
          schema: { type: number, format: double }
        - name: maxY
          in: query
          required: false
          schema: { type: number, format: double }
        - name: sortBy
          in: query
          required: false
          schema: { type: string, default: x, enum: [ x, y ] } # Добавьте поля для сортировки
          description: Поле для сортировки
        - name: direction
          in: query
          required: false
          schema: { type: string, default: asc, enum: [ asc, desc ] }
          description: Направление сортировки
      responses:
        '200':
          description: Список точек
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointResponse'
  /functions:
    post:
      summary: Создать новую табулированную функцию
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionCreateRequest'
      responses:
        '201':
          description: Функция создана
          headers:
            Location:
              schema:
                type: string
                example: /api/functions/123
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
...
  /functions/{id}/operations/binary:
    post:
      summary: Выполнить бинарную операцию над двумя функциями
      description: |
        Выполняет одну из операций (+, -, *, /) между функцией `id` и второй функцией `secondFunctionId`.
        Результат может быть сохранён как новая функция или возвращён только в ответе (в зависимости от политики сервера).
      tags: [Operations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BinaryOperationRequest'
      responses:
        '200':
          description: Операция успешно выполнена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '400':
          description: Некорректные входные данные или несовместимые функции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Одна из функций не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/{id}/derive:
    post:
      summary: Вычислить производную табулированной функции
      description: |
        Вычисляет n‑ую производную (по умолчанию первую) табулированной функции с использованием численного дифференцирования.
      tags: [Operations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DerivativeRequest'
      responses:
        '200':
          description: Производная успешно вычислена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '404':
          description: Функция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    UserResponse: # Соответствует UserResponseDTO.java
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
        username:
          type: string
          description: Имя пользователя
        role:
          type: string
          enum: [ USER, ADMIN, MODERATOR ] # На основе org.example.entity.Role
          description: Роль пользователя
        createdAt:
          type: string
          format: date-time
          description: Дата создания аккаунта
    FunctionListResponse: # Соответствует TabulatedFunctionListDTO.java
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID функции
        name:
          type: string
          description: Имя функции
        createdAt:
          type: string
          format: date-time
          description: Дата создания функции
        userId:
          type: integer
          format: int64
          description: ID владельца
        username:
          type: string
          description: Имя владельца
        pointsCount:
          type: integer
          minimum: 0
          description: Количество точек в функции
    PointResponse: # Соответствует PointDTO.java / PointResponseDTO.java
      type: object
      required: [ x, y, tabulatedFunctionId ]
      properties:
        id:
          type: integer
          format: int64
          description: ID точки
        x:
          type: number
          format: double
          description: Координата X
        y:
          type: number
          format: double
          description: Координата Y
        tabulatedFunctionId:
          type: integer
          format: int64
          description: ID функции, которой принадлежит точка
    FunctionCreateRequest:
      type: object
      required: [ownerId, name, points]
      properties:
        ownerId:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: "my_quadratic"
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
...
    FunctionResponse:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PointData:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
          format: double
          example: 1.0
        y:
          type: number
          format: double
          example: 2.0
    BinaryOperationRequest:
      type: object
      required: [secondFunctionId, operation]
      properties:
        secondFunctionId:
          type: integer
          description: ID второй функции, над которой выполняется операция
        operation:
          type: string
          description: Тип бинарной операции
          enum: [ADD, SUB, MULT, DIV]
    DerivativeRequest:
      type: object
      properties:
        order:
          type: integer
          description: Порядок производной (1 — первая, 2 — вторая и т.д.)
          minimum: 1
          default: 1
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationFailed"
        message:
          type: string
          example: "Поле 'name' не может быть пустым"
        field:
          type: string
          example: "name"
        timestamp:
          type: string
          format: date-time